refactor per feedback
